oppdaterer hjemmesalg aktiviteten
