+++ app/src/main/java/com/julia/bachelor/Main.java
@@ -47,7 +47,7 @@ public class Main extends Activity
         // update the main content by replacing fragments
         FragmentManager fragmentManager = getFragmentManager();
         fragmentManager.beginTransaction()
-                .replace(R.id.container, Hovedside.newInstance(position + 1))
+                .replace(R.id.container, RapportFragment.newInstance(position + 1))
                 .commit();
     }
 
@@ -62,7 +62,7 @@ public class Main extends Activity
                 bundle = new Bundle();
                 bundle.putSerializable("beholdning", Beholdning);
                 bundle.putSerializable("salg",BeholdningUt);
-                RapportFragment myf = RapportFragment.newInstance(1);
+                HovedsideFragment myf = HovedsideFragment.newInstance(1);
                 myf.setArguments(bundle);
                 FragmentTransaction fragmentTransaction = getFragmentManager().beginTransaction();
                 fragmentTransaction.replace(R.id.container, myf);
