+++ app/src/main/res/layout/rapport_fragment.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".Hovedside">
+    tools:context=".RapportFragment">
 
     <ScrollView
         android:layout_width="match_parent"
